Fixed bugs on polyline editing regarding edge cases
This commit fixes couple bugs. One of them is polylines (or polygons)
can be  reduced to a single point by clicking on markers to remove
them. Second, if an edge point is clicked for a polyline (not
applicable to polygons since they are a circular linked list) code
produces a js error since previous or next might not be present.