Add support for custom units in scale control
Custom unit works via a callback that is passed the current
maximum scale in meters and supposed to return a hash containing
a caption (number and unit) to be displayed in the scale control
and a ratio (rounded meters / max meters) for correct sizing of
the scale control.

For convenience and consistency, the callback also receives a
reference to the default leaflet scale control rounding function.

Styles have been more made slightly more generic to (theoretically)
support an infinite number of scales (treating the first and last
ones as special).

The controls example has be amended with an example of a custom
scale (nautical miles).