Allow leaflet to ignore DOM events for some nodes.
This change proposes a possible way to allow elements contained
inside the DOM structure of a leaflet map to mark themselves
in such a way that all events inside them are ignored by
leaflet and passed through to the other handlers in the system.

I believe this is needed for proper integration with frameworks
that connect to events at the document level.