Add a test that shows the DOM change failure
The icon should switch from blue to red and back, but does not in IE
because the DOMelement has disappeared, as in issue #561. Instead, the
mouseover event is fired on any mouse motion in the marker. In addition,
in FF & Chrome the mouseover and mouseout events are continously fired
on any movement, which gives the correct visual results but causes
a lot of superflous event handling.