Issue where the tile layer is not reset and updated when redraw is called.
I was able to replicate this issue: https://github.com/Leaflet/Leaflet/issues/1797. The problem is that redraw does not call update and reset on the TileLayer, therefore causing the above issue. I am proposing simply doing it before calling drawTile. I built and tested it, and it does fix the issue.