Fix IE pointer events passing through to the map
Pointer events in IE would trigger on the map even when not started on the map. For example, if you had a dialog box on top of the map and dragged inside the map, the map would move. I believe this is because the handler never gets added to the obj[eventsKey] map. I noticed in the corresponding _off function below that the logic reaches there because of the multiple if/else statements. So I modified the _on even to match and it seems to fix the problem.