Update GeoJSON.js
I modified the function to explicitly return the geoJSON properties. When using geoJSON's with different CRS's this function used to return the original CRS name but with geometry converted to WGS84 lat/lng format.  If the data was reused for another projection the locations would be in the wrong locations as the coordinates didn't match the CRS.  Now this function just returns the lat/lng but without the CRS attached.