Added zoomFactor
Added the ability to control the zoom factor via Control.Zoom.js and ScrollWheel zooming.

This is in regards to fractional zooming. It will zoom the map in increments of whatever the zoomFactor is whenever you use the Zoom Control or Scrollwheel.

Pass the `zoomFactor` in the map options. Default value is 1. You can change the map's zoom factor at any time by using `map.zoomFactor = 0.25`.

I added a `zoom-factor.html` debug file to experiment with.

Also @mourner if you want, we can change the name zoomFactor to something else. Maybe zoomStep or zoomIncrement? zoomFactor sounded most straightforward at the time. This is a minor feature but it really should be in there in order to make fractional zooming shine.

Signed-off-by: Jake Wilson <jake.e.wilson@gmail.com>