Allow to use console in tests
This prevents us from adding /* eslint no-console: 0 */ each time
we need to use console.log/trace for debugging.

This also means that we will not be blocked if we forget a console
in the tests before commiting, but this should also happen while using
the inline "no-console" config, and given that this only affects tests
files, I think the confort of being able to use console easily worths
the risk of pushing a console in the tests files.