trigger tile compositing in Safari
This kind of reverts #2377 that doesn’t make sense after tile animation
refactorings, because it makes Safari terribly slow. It reintroduces
the tile gaps, but oh well, we choose the lesser evil here.

Also, performance in latest iOS is the same with and without the patch
so there’s no harm in removing this.