updating .getLayers() to accept an array of ids.
While .getLayer(<layer id>) returns a specific layer, the getLayers() function does not accept an argument to get a list of matching layers.  This change modifies the .getLayers() function so that an array of layer ids may be passed into the function, and it will return the matching layers.  This change is backwards compatible, if no array is passed to function it will still return an array with all of the layers in the group.