Add shiftkey option to scroll wheel zoom
This PR adds the possibility to set the map option "scrollWheelZoom" to enable scroll wheel zoom only when the shift key is pressed. Useful when you want to disable scroll wheel zoom but still make it possible to use the scroll wheel in some way. 

I'm not sure if this breaks any other mouse/keyboard events so it would be great if someone with better Leaflet knowledge could take a look at it.