Fix GridLayer pixelBounds computation
Map size does not change from one zoom to another, so we dont need
to scale it.

Without this change, we are creating more or less 4x tiles when
zooming in.
There are still cases where we are requesting a bit more tiles than
what I would expect, but I see it also happens on 0.7.3