Cleaning up code and using L.DomUtil.create instead of document.createElement
Using `L.DomUtil.create()` to simplify this function.  Instead of using `document.create`, className, and appendChild function, we are creating the element, defining its class, and immediately appending it to its parent before setting custom parameters.  Additionally i moved all variable creation to the top of the function, and put the variables in order of how they are appended to improve legibility.