New layer method toSingleLayersArray
to always get an array of *non-group* (single) layers, whether we deal with a single layer already or with a Layer Group.
By having a generic method at the Layer base class level, we can use it on any layer we receive as argument, without having to re-implement the Layer Group detection and acting accordingly.