Add ColorIcon - a traditional marker icon of configurable color
Icon is implemented using data: URL to generate SVG image of requested
color. The shadow is embedded PNG file of original marker.

Thanks to use of data: URLs this icon does not use any external files.

The SVG image of the marker is the original Leaflet marker without
gradient and some redundant XML.