Use translate() in DomUtil.setTransform
Use translate() instead of translate3d() because the z transformation
is 0 anyway, and browsers seem to better support the former
transformation (i.e. the interpolation (e.g. when upscaling) remains
bicubic (in browsers that support it, e.g. Chromium)  instead of
bilinear as noticeable (in Chromium) in https://jsfiddle.net/Lb6abxmu/)