Normalize line endings to UNIX-style
There is currently a mix of Windows-style (CR-LF) and Unix-style (LF)
line endings in the Leaflet repository.

This PR makes all text files consistently use Unix-style (LF) line
endings.

The inconsistent EOL-style is particularly problematic for the file
/dist/leaflet-src.js, constructed from different source files, and
which produces a mix of CRLF and LF line endings within the same file.