Escape html elements inside code sample (#5432)
The markup renderer will still literally render html tags inside "pre" and "code" unless they're escaped. The result was that the example code would be rendered as:
var baseMaps = {
	"Grayscale": grayscale,
	"Streets": streets
};
rather than
var baseMaps = {
	"<span style='color: gray'>Grayscale</span>": grayscale,
	"Streets": streets
};
because the "Grayscale" text would be literally styled as a span with gray color, and then recolored by the syntax highlighter.

Escaping the span tags fixes this.