Feat(IconDefault): read all options from CSS
so that build engines that rewrite URL's no longer break Leaflet icons image path detection and its previously hard-coded file names.
Note that CSS margins (to specify anchors position) are negative, whereas anchors are specified as positive in JS. This is to be closer to when positioning a custom icon through CSS (instead of specifying its iconSize and iconAnchor), where we have to use negative margins to offset the icon image to the top left of the icon tip.
Instead of hard coding options in L.Icon.Default.prototype, they are now applied **per** instance.
With the new option `classNamePrefix`, it also becomes possible to use this new initialization behaviour to generate more Icons from CSS properties, although I am not sure there is a real use case for it, because it is easier to define Icons in JS, and the app developer can easily get a dynamically generated image url if he/she can code JS anyway (e.g. in webpack: `iconUrl: require('icon.png')`). Only the IconDefault was troublesome because it previously required the app developer to tweak it if the build engine fiddled with the urls, instead of working by itself no matter what.
There is a slight change in usage of `imagePath` however: it is used only once at first CSS retrieval (typically when the first marker is added to a map). Modifying it later on no longer has any effect, whereas previously it would change the resulting iconUrl for any icon redraw.