Fix(IconDefault): use CSS cursor for FF colours override
as mentioned in https://github.com/Leaflet/Leaflet/issues/4968#issuecomment-328277840 when Firefox high contrast mode / colours override setting is set to "Always", the `background-image` CSS rule is removed, which breaks the path detection.
Using `cursor` instead makes it immune to that setting, although the resulting class looks less "elegant", further from a style for a marker. But anyway this class is used only to communicate data to the JS, not to be directly applied to a rendered element.