Fix(IconDefault): use background-image first, fallback to cursor
as mentioned in https://github.com/Leaflet/Leaflet/pull/5771#issuecomment-333618629, IE treats URL's in `cursor` CSS rule as relative to the HTML page, instead of relative to the CSS file, as expected by Leaflet. Therefore it is unusable for IE.
However, as mentioned in https://github.com/Leaflet/Leaflet/issues/4968#issuecomment-328277840, Firefox may override the `background-image` CSS rule (in case high contrast / colours override setting is set to "always"). In order to fix this corner case, we still need to fallback to `cursor`.
Also added a dedicated test spec for this fallback mechanism.