Refactor(IconDefaultSpec): no need to replace Marker.prototype.options.icon
just using a new instance of L.Icon.Default is enough to test its behaviour, without having to interfere with the instance in Marker.prototype.options.icon.
However, we still need to properly remove the default options on L.Icon.Default (i.e. on its prototype), so that next tests that may reload Leaflet (hence which re-instantiate a default icon) are not affected.