Docs(MapPanes): explicitly list layer types in overlayPane
"overlayPane" is the default pane for Layers (https://github.com/Leaflet/Leaflet/blob/v1.2.0/src/layer/Layer.js#L36), but it is overridden for many sub-classes (Marker, DivOverlay, GridLayer).
It would be easier to understand where things go if we explicitly list them in Map Panes section as well. It is almost the case, except for Image Overlay and Video Overlay.