Refactor(IconDefault): strip path from iconUrl if imagePath is provided
so that we maintain backwards compatibility with current situation in the case both `imagePath` and `iconUrl` (and other URL) options are provided by the application.
In this case, `imagePath` replaces the entire path of the CSS provided URL for all icons and shadows.
For app provided `iconUrl` option (same for shadow and Retina), no stripping happens. Therefore, they could even include a partial path before the file name. It will be concatenated with `imagePath`.
With this new version, IconDefault emulates much closer the current behaviour: if `imagePath` option is already provided, it is used as entire path, and only file name is automatically provided by IconDefault (except that previously it was hard-coded in IconDefault.prototype.options, now it is retrieved from leaflet.css).
If `iconUrl` is also provided, it is directly appended after `imagePath` as-is.