docs(GeoJSON): use RFC link for previous HTML docs
RFC 7946 was officially published in August 2016, i.e. before Leaflet version 1.0.0 (Sep 2016). The previous GeoJSON spec website (geojson.org) displays a big "obsolete" disclaimer on it, making it confusing for developers which version they should refer to. Replace the link to use the RFC that was in use at that time, on all HTML docs since it was available, i.e. from Leaflet 1.0.0.