Add ceilZoom to GridLayer to select zoom level used between integers
By default, a fractional zoom level is rounded, that is, the tiles
switch zoom level at halfway point which creates undesirable >1
scale when a canvas is used within a GridLayer tile. This change
allows avoiding >1 scales by setting ceilZoom to true that makes
the tile zoom level to be always larger than a non-integer zoom
level.