Reimplement DoubleTap simulation in more simple and reliable way
Base handler on 'click' event instead of 'touch*' / 'pointer*'.
Ensure that simulated 'dblclicks' are fired only when needed.

ATM it's safe to use in any known browser, with or without touch support,
and it is equally efficient with or without pointer events.

Additionally: reduce `delay` value to 200, as it is closer to actually observed.