fix tooltipAnchor behavior for different directions (#7155)
This commit fixes issue #6764, see the discussion examples there.
It's an alternative implementation for pull request #6116.

It makes `tooltipAnchor` behave the same like `popupAnchor` regardless of the tooltip's direction. With one exception: the `auto` direction flips the `x` axis of `tooltipAnchor` and the tooltip's `offset` when it switches to `left` depending on the position on the screen. `auto` therefore assumes that the icon is somewhat symmetrical to the left and the right of the y axis, which is the case for the default icon.