Get rid of some legacy IE-related pointer quirks. (#7195)
That code was initially intended to support mspointer-touch, but currently
same code paths are used in any PointerEvent-enabled browser, and may cause
real issues (e.g. that broke Leaflet.draw).

Note:
It's known that preventDefault is needed for correct processing of touch in IE10.
But I haven't found any evidence that it is required for IE11, so I've removed some
excessive code.
It's blind shot as I have no IE10-11 devices to test, but it will be easy to fix on request.