Quote strings in url()
I'm having problems with a CSS postprocessor. It unwraps the CSS file and adds a hash to each of the referred images to assure that those images will be reloaded after a change. That hash value is appended using a ?somehashvalue which causes trouble when loading.
By quoting the urls in CSS, this problem can easily be fixed.