epic update of all demo URLs in the documentation
This is an epic update of almost every URL in the plugin list
of the Leaflet documentation.

I've manually tested almost every plugin in almost every repository
and opened lot of issues in each repository trying to investigate
the upstream status.

If something is not in order we invite subsequent pull requests.

Thank you Leaflet developers for all that you do!