test: add coverage for Layer\_addZoomLimit
Resolves #7762

Checks that zoom levels are updated to non-NaN values when
minZoom and maxZoom are set to non-NaN values in one or many
layers that are passed into addZoomLimit.

Checks that min zoom level is updated to NaN when minZoom is
set to NaN in at least one layer that passed into addZoomLimit,
but maxZoom is never set in any layer. Vice versa for when
maxZoom is set to NaN in at least one layer but minZoom is never
set in any layer. Not sure if this is a feature or a bug, so
need to followup with reviewers.

Checks that zoom levels are undefined when minZoom and maxZoom
are both set to NaN in a layer that is passed into addZoomLimit.