DomEvent.DoubleTap ignores clicks on <label>s
Fixes #8225.

When clicking on a <label> linked to an <input> (or vice
versa), browsers trigger an extra click on the related
element. This causes the DoubleTap logic to activate
and fire a ghost dblclick event.

This change makes the DoubleTap logic ignore clicks on
labels or children of labels (but not when there's
an <input> on the propagation path of the event, which
means there's an <input> nested inside a <label>).