Quick-start: fix link in code block.
The quickstart guide contains a code block with a link;
this should be rendered as an `<a>` HTML tag, but is
rendered as an actual link. I've fixed the Markdown by
changing it to a fenced code block. For consistency,
I've also adjusted the other code blocks in the page.