Update download.md
I included an overview of the different types of versions that exist for Leaflet. This will allow the user to view the options quickly and easily. Also, I included an html overview section which outlines the html structure on how to incorporate Leaflet's hosted version within the header of the html file. I also included other aspects like the user's index.js and css. The goal is to show all together how the main setup should look like to allow the user to get a good base when first starting. Lastly, I referenced the quick start guide to direct the user just in case they have more intricate details about starting.