More reliable PopupSpec's keepInView tests
The previous test was working as a side-effect of how far it was moving. If it were movnig a short distance that still triggered a pan, the test would fail because it would have invoked animations and the map would not have immediately updated position.

These tests should work reliably whether we're moving a short or long distance, so I have added a new test to cover short distances separately. I have also updated both tests to wait for the 'moveend' event, so they will continue to reliably work no matter what happens to `Map.panBy` logic in the future.

After this change, the average combined runtime from both tests is 271ms. Compared to the original test, this remains a speed up from 812ms to 271ms (3x).