Do not propagate click when element has been removed from dom
This fixes an edge case where, when the initial click target
element is removed (or one of its parents) during the click
listener processing, then that click will not be disabled,
and thus transmited to the map.

A concrete exemple would be a popup with a clickable button
that would replace the popup content on click. In this case,
the click event will be transmited to the map, which will
then close the popup (as for any normal click on the map when
a popup is open).